requirements.txt
pandas==1.4.0
pyTelegramBotAPI==4.5.0
dropbox==11.30.0
PyGithub==1.55